requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.13.2

# Tools
django-extensions==3.0.5

# Testing
mypy==0.782
pytest==6.0.1
pytest-sugar==0.9.4
pytest-django==3.4.4
factory-boy==2.12.0
pylint-django==2.3.0
importlib-resources==3.0.0
pyfakefs==4.1.0

# Code quality
flake8==3.8.3
